Bug: Incorrect centroid mapping and dtype handling in feature flattening

This PR fixes a bug where centroid_x and centroid_y were incorrectly mapped from (row, col) tuples returned by regionprops.

+++ src/playnano/analysis/modules/feature_detection.py
@@ -215,7 +215,7 @@ class FeatureDetectionModule(AnalysisModule):
                 {
                     "frame_timestamp": frame_ts,
                     "label": int(prop.label),
-                    "area": int(prop.area),
+                    "area": float(prop.area),
                     "min": float(vals.min()),
                     "max": float(vals.max()),
                     "mean": float(vals.mean()),
